Rename "core" directory to "chess"
Bench: 2266395
