+++ src/chess/castling_rights.h
