+++ src/chess/magic.h
