+++ src/network/nnue.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "../external/incbin/incbin.h"
-#include "../core/constants.h"
+#include "../chess/constants.h"
 #include "../utils/logger.h"
 #include "activations/crelu.h"
 #include "layers/accumulator.h"
