+++ src/search/search_limits.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "../core/constants.h"
+#include "../chess/constants.h"
 
 #include <optional>
